Added support for uninit_bg
Added support for the uninit_bg feature.  This feature actually enables
the EXT4_FEATURE_RO_COMPAT_GDT_CSUM feature bit and makes use of the
uninit bg flags.  When enabled, EXT2_BG_INODE_UNINIT causes reading of
the inode allocation bitmap for that group to be skipped and assumed to
be all zero.  Also the EXT2_BG_BLOCK_UNINIT flag causes the same for the
block allocation bitmap.  This flag is set or cleared when writing the
block groups back dependong on whether or not any blocks are allocated
in that group.  When set the block bitmap for that group is not written.