Fixed "musical blocks" bug
Block positions would change on repeated invocations rather than being
deterministic.  This was caused by inodes being optimised in an order
decided by their average position on disk.  When a block group
overflowed, the latter files would have blocks allocated elsewhere,
possibly in earlier groups, which would bring their average position
down, which would cause them to be optimised earlier on the next
defrag.

Fixed this by changing the way that the average position is calculated.
Now when blocks are found outside the native group of the inode, its
average position is forced to INFINITY in float form, which is converted
to an integer as ~0, thus forcing the sort order of the inode to always
follow those inodes which are in the native group, and all inodes that
overflowed the native group to maintain constant sort order relative to
each other, which is their inode number order.