Allow block allocator to force blocks left when prioritized
Changed block allocator to force blocks to the left instead
of allocating from their native block group when they have
been given a negative priority in the priority file.  This
allows for more of the intended effect when using the
priority file since one specifies a negative priority to
force files to the left.  Previously doing so only forced
them to the left within their native block group, until
that group ran out of room.