Add ext4 extents support
First draft of code to parse and update extent trees.
Still can not modify the depth of the tree, so can fail
if a file becomes more fragmented than it was, and will
not take advantage of the fact that if the number of
extents is reduced to 4 or less they can be moved back
into the inode instead of the external tree.