O_DIRECT, readv and hash table optimizations
Reorganized the buffer pool into separate buffer head
structures with pointers to the data buffers, rather than
having the buffer follow the header structure.  This
allowed aligning the actual buffers to satisfy the O_DIRECT
buffer alignment requirements.  Enable O_DIRECT with fcntl
during buffer relocations to reduce cpu overhead moving
data in and out of buffers.

Modified buffer read/write code to add the reqest to an
iovec queue to batch them together and send them down to
the kernel as a group with readv().

Increased the size of the hash table which lists buffers
hashed by the low order bits of the block number from
1024 entries to 16k entries to greatly reduce collisions
with larger buffer pools.