Vastly imrpoved performance when forcing
When too many buffers are used for rescue, some need to be freed.  First
defrag tries to migrate buffers whose destinations are free to the output
pool, causing them to be written out of order.  If that does not free
enough, some buffers are forced out to a temporary location at the end
of the disk.  This forcing was horribly slow because it was done one
block at a time in decending order, rather than ascending order, which
prevented the new O_DIRECT queue code from coalescing the writes.  It
also did not show the temporary writes on the screen map.

I created a new FORCE buffer type and changed the force code to convert
a buffer to this type and let it be flushed in the normal way which
correctly updates the display map and coalesces the writes for speed.
This has resulted in significantly faster defrags on an fs that has
not been recently defragged already, where forces happen frequently.
My test case went from almost 14 to almost 7 minutes when defrag was
given 128mb of buffers and from almost 9 to almost 7 minutes when it
was given 512mb of buffers.  During the force before the change, the
IO throughput would drop to ~1-2mb/s and defrag would appear to hang
for a long period of time performing a write pattern that was actually
the migrations before the force.  Now the force shows up properly
as writes to the end of the disk, and throughput is normal.