Added new block map code
Defrag builds a translation matrix that maps every disk block's
current location to its new location after defrag, and vice versa.
Until now that has been done using a pair of arrays that contain
the destination/source block number.  This uses 8 bytes of ram
for each block on the disk, which does not scale to modern large
disk sizes, where a 320 gb disk requires 600+ mb of ram just for
the translation matrix.

I am replacing the old arrays with a new extent based data
structure that is double indexed in a pair of glib balanced
binary trees.  This way blocks that are mapped from a set
of consecutive source addresses to a set of consecutive
destination addresses will only need one extent to represent
them all.  Blocks which are free will require no storage at
all.  This is the initial implementation of the helper
functions to implement the translation map.  A following
patch will modify defrag to use the new map.