Removed inode average location sorting
Defrag used to compute an average block position for
each inode and use that to sort the optimization order.
It did this to minimize the distance blocks are moved
by keeping files in roughly the same order they started
in even if the blocks of higher inode numbers tend to
be located before lower number inodes.

This required 4 bytes of ram for every possible inode
in the fs, which amounted to ~250 mb of ram on a 1 TB
fs.  I don't think that this actually ends up doing much
good and is a waste of ram, so I have removed it.

Also added dump2inodes.py script to generate an inode
priority file from ureadahead --dump.