Remove many time() syscalls
When scanning inodes, time() was being called between each inode to figure
out if enough real time had passed to force a screen update.  The screen
is also updated every 2% of the total inodes, so this isn't really needed
and on larger disks with many inodes, this added up to a fair amount of
wasted cpu time.