Fix buffer read count
The buffer read count was exceeding the buffer write count by
the number of forced buffers.  This was because when a force
happens, the block is written to a temporary location, and
then read back in later, before finally being written to the
true destination.  The forced write was not counted as a normal
write, but the both reads were being counted.  Corrected by
decrementing the read count when writing a force buffer, thus
undoing the first time it was counted, and it will be counted
as a normal read again when it is read later.