Improved the progress indicator
The stat line used to indicate the progress of the destination cursor
as the relocation loop walked the disk, and claimed to be relocating
xx MB where that size simply corresponded to the destination cursor.
This was misleading, so the relocation loop has been changed to count
the number of blocks it is going to move and print that at the start,
then print the progress as the number of MB actually moved, and as
a percentage of the total to be moved.